Moved MiniOzoneClusterProvider to its own reusable class
