+++ hadoop-ozone/integration-test/src/test/java/org/apache/hadoop/ozone/scm/node/TestDecommissionAndMaintenance.java
@@ -36,10 +36,10 @@ import org.apache.hadoop.hdds.scm.node.states.NodeNotFoundException;
 import org.apache.hadoop.hdds.scm.pipeline.Pipeline;
 import org.apache.hadoop.hdds.scm.pipeline.PipelineManager;
 import org.apache.hadoop.hdds.scm.server.StorageContainerManager;
+import org.apache.hadoop.ozone.MiniOzoneClusterProvider;
 import org.apache.hadoop.ozone.MiniOzoneCluster;
 import org.apache.hadoop.ozone.TestDataUtil;
 import org.apache.hadoop.ozone.client.OzoneBucket;
-import org.apache.hadoop.ozone.om.OMConfigKeys;
 import org.apache.ozone.test.GenericTestUtils;
 import org.junit.After;
 import org.junit.AfterClass;
@@ -56,9 +56,6 @@ import java.util.Arrays;
 import java.util.Iterator;
 import java.util.List;
 import java.util.Set;
-import java.util.UUID;
-import java.util.concurrent.ArrayBlockingQueue;
-import java.util.concurrent.BlockingQueue;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.TimeoutException;
 import java.util.stream.Collectors;
@@ -104,135 +101,7 @@ public class TestDecommissionAndMaintenance {
 
   private ContainerOperationClient scmClient;
 
-  private static MiniClusterProvider clusterProvider;
-
-  /**
-   * Class to create mini-clusters in the background.
-   */
-  public static class MiniClusterProvider {
-
-    private int preCreatedLimit = 1;
-    private boolean shutdown = false;
-
-    private final OzoneConfiguration conf;
-    private final MiniOzoneCluster.Builder builder;
-    private Thread createThread;
-    private Thread reapThread;
-
-    private final BlockingQueue<MiniOzoneCluster> clusters
-        = new ArrayBlockingQueue<>(preCreatedLimit);
-    private final BlockingQueue<MiniOzoneCluster> expiredClusters
-        = new ArrayBlockingQueue<>(1024);
-
-    public MiniClusterProvider(OzoneConfiguration conf,
-        MiniOzoneCluster.Builder builder) {
-      this.conf = conf;
-      this.builder = builder;
-      createThread = createClusters();
-      reapThread = reapClusters();
-    }
-
-    public synchronized MiniOzoneCluster provide()
-        throws InterruptedException, IOException {
-      ensureNotShutdown();
-      return clusters.poll(100, SECONDS);
-    }
-
-    public synchronized void destroy(MiniOzoneCluster c)
-        throws InterruptedException, IOException {
-      ensureNotShutdown();
-      expiredClusters.put(c);
-    }
-
-    public synchronized void shutdown() throws InterruptedException {
-      createThread.interrupt();
-      createThread.join();
-      destroyRemainingClusters();
-      reapThread.interrupt();
-      reapThread.join();
-      shutdown = true;
-    }
-
-    private void ensureNotShutdown() throws IOException {
-      if (shutdown) {
-        throw new IOException("The mini-cluster provider is shutdown");
-      }
-    }
-
-    private Thread reapClusters() {
-      Thread t = new Thread(() -> {
-        boolean shouldRun = true;
-        while(shouldRun || !expiredClusters.isEmpty()) {
-          try {
-            if (Thread.interrupted()) {
-              throw new InterruptedException();
-            }
-            MiniOzoneCluster c = expiredClusters.take();
-            c.shutdown();
-          } catch (InterruptedException e) {
-            shouldRun = false;
-          }
-        }
-      });
-      t.setName("Mini-Cluster-Provider-Reap");
-      t.start();
-      return t;
-    }
-
-    private Thread createClusters() {
-      Thread t = new Thread(() -> {
-        while (!Thread.interrupted()) {
-          MiniOzoneCluster cluster = null;
-          try {
-            builder.setClusterId(UUID.randomUUID().toString());
-
-            OzoneConfiguration newConf = new OzoneConfiguration(conf);
-            List<Integer> portList = getFreePortList(4);
-            newConf.set(OMConfigKeys.OZONE_OM_ADDRESS_KEY,
-                "127.0.0.1:" + portList.get(0));
-            newConf.set(OMConfigKeys.OZONE_OM_HTTP_ADDRESS_KEY,
-                "127.0.0.1:" + portList.get(1));
-            newConf.set(OMConfigKeys.OZONE_OM_HTTPS_ADDRESS_KEY,
-                "127.0.0.1:" + portList.get(2));
-            newConf.setInt(OMConfigKeys.OZONE_OM_RATIS_PORT_KEY,
-                portList.get(3));
-            builder.setConf(newConf);
-
-            cluster = builder.build();
-            cluster.waitForClusterToBeReady();
-            clusters.put(cluster);
-          } catch (InterruptedException e) {
-            if (cluster != null) {
-              cluster.shutdown();
-            }
-            break;
-          } catch (IOException | TimeoutException e) {
-            throw new RuntimeException("Unable to build cluster", e);
-          }
-        }
-      });
-      t.setName("Mini-Cluster-Provider-Create");
-      t.start();
-      return t;
-    }
-
-    private void destroyRemainingClusters() throws InterruptedException {
-      while(!clusters.isEmpty()) {
-        try {
-          expiredClusters.add(clusters.poll(100, TimeUnit.MILLISECONDS));
-        } catch (InterruptedException e) {
-          // Do nothing
-        }
-      }
-    }
-
-    private List<Integer> getFreePortList(int size) {
-      return org.apache.ratis.util.NetUtils.createLocalServerAddress(size)
-          .stream()
-          .map(inetSocketAddress -> inetSocketAddress.getPort())
-          .collect(Collectors.toList());
-    }
-  }
+  private static MiniOzoneClusterProvider clusterProvider;
 
   @BeforeClass
   public static void init() {
@@ -260,7 +129,7 @@ public class TestDecommissionAndMaintenance {
     MiniOzoneCluster.Builder builder = MiniOzoneCluster.newBuilder(conf)
         .setNumDatanodes(numOfDatanodes);
 
-    clusterProvider = new MiniClusterProvider(conf, builder);
+    clusterProvider = new MiniOzoneClusterProvider(conf, builder);
   }
 
   @AfterClass
